+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 multiline Python logs (like stack traces)
 being split into multiple CloudWatch log entries.
 
+### Changed
+- Log internal SDK errors using
+`warning` level instead of `exception`.
 
 ## [1.4.0-rc.1] - 2021-02-15